+++ src/backend/resolvers/queries/organizations/mbm/lbm/joins/JOIN_lbms_payers.ts
@@ -1,0 +1,5 @@
+const JOIN_lbms_payers = (parent, args, { pulseCoreDb }) => {
+  return pulseCoreDb.collection('JOIN_lbms_payers').find().toArray()
+}
+
+export default JOIN_lbms_payers