+++ src/backend/typeDefs/mutations/organization/mbm/lbm/relational-typeDefs/index.ts
@@ -3,6 +3,7 @@
 import lbmAndLbmTypeTypeDefs from './lbmAndLbmType'
 // import lbmAndPersonTypeDefs from './lbmAndPerson'
 // import lbmAndKeyEvent from './lbmAndKeyEvent'
+import lbmAndPayer from './lbmAndPayer'
 
 export default [
   // ...serviceAndServiceCategoryTypeDefs,
@@ -10,4 +11,5 @@ export default [
   ...lbmAndLbmTypeTypeDefs,
   // ...lbmAndPersonTypeDefs,
   // ...lbmAndKeyEvent,
+  ...lbmAndPayer,
 ]