+++ src/frontend/components/BusinessObjectModal/shared/widget/mbmPayerConnectionComponents/ConnectionPanel/ButtonCluster.js
@@ -8,29 +8,10 @@ import styled from '@emotion/styled'
 
 import Button from 'frontend/components/Button'
 import Spinner from 'frontend/components/Spinner'
-import { GET_EVENTS, GET_JOIN_PATHWAYS_AND_PEOPLE } from 'frontend/api/queries'
 
 import Color from 'frontend/utils/color'
 import Spacing from 'frontend/utils/spacing'
-import stripTypename from '../../../../../Orion/shared/strip-typename'
-
-// TODO: Pass in mutations and refetch gql tags to button cluster
-import {
-  UPSERT_PATHWAYS_AND_PERSON_CONNECTION,
-  DELETE_PATHWAYS_AND_PERSON_CONNECTION,
-} from 'frontend/api/mutations'
-
-const REQUIRED_FIELD_KEYS = [
-  'pathwaysId',
-  'personId',
-  'indicationIds',
-  'pathwaysInfluencerTypes',
-  'pathwaysInfluencerTypes',
-  'position',
-]
-
-const areRequiredFieldsBlank = (data) =>
-  REQUIRED_FIELD_KEYS.some((key) => _.isEmpty(data[key]))
+import stripTypename from 'frontend/Orion/shared/strip-typename'
 
 const Alert = (props) => {
   return <MuiAlert elevation={6} variant="filled" {...props} />
@@ -54,70 +35,31 @@ const ButtonsWrapper = styled.div({
 const ButtonCluster = ({
   isNewConnectionBeingCreated,
   cancelHandler,
-  connectionData,
+  stagedConnection,
   setWhetherUnsavedChanges,
   setWhetherNewConnectionBeingCreated,
-  changeConnection,
-  connectionsData,
+  selectConnectionId,
+  connections,
+  refetchQueries,
+  mutationDocs,
 }) => {
   const [snackbarOpen, toggleSnackbar] = useState(false)
 
-  connectionData = stripTypename(_.cloneDeep(connectionData))
-
-  const {
-    _id,
-    pathwaysId,
-    personId,
-    indicationIds,
-    pathwaysInfluencerTypes,
-    tumorTypeSpecialty,
-    internalFields,
-    position,
-    priority,
-    alert,
-    exclusionSettings,
-    startDate,
-    endDate,
-    startQuarter,
-    endQuarter,
-  } = connectionData
-
-  const dataToPersist = {
-    _id,
-    pathwaysId,
-    personId,
-    indicationIds,
-    pathwaysInfluencerTypes,
-    tumorTypeSpecialty,
-    internalFields,
-    position,
-    priority,
-    alert,
-    exclusionSettings,
-    startDate,
-    endDate,
-    startQuarter,
-    endQuarter,
-  }
+  stagedConnection = stripTypename(_.cloneDeep(stagedConnection))
 
   const [upsert, { loading: upsertLoading }] = useMutation(
-    UPSERT_PATHWAYS_AND_PERSON_CONNECTION,
+    mutationDocs.upsert,
     {
       variables: {
-        input: dataToPersist,
+        input: stagedConnection,
       },
-      refetchQueries: [
-        { query: GET_EVENTS },
-        { query: GET_JOIN_PATHWAYS_AND_PEOPLE },
-      ],
+      refetchQueries,
       awaitRefetchQueries: true,
       onCompleted: (res) => {
         if (isNewConnectionBeingCreated) {
           setWhetherNewConnectionBeingCreated(false)
           const newConnectionId = Object.values(res)[0]._id
-          changeConnection(
-            connectionsData.find(({ _id }) => _id === newConnectionId)
-          )
+          selectConnectionId(newConnectionId)
         }
 
         toggleSnackbar(true)
@@ -127,24 +69,18 @@ const ButtonCluster = ({
     }
   )
 
-  const [deleteConnection] = useMutation(
-    DELETE_PATHWAYS_AND_PERSON_CONNECTION,
-    {
-      variables: {
-        input: dataToPersist,
-      },
-      refetchQueries: [
-        { query: GET_EVENTS },
-        { query: GET_JOIN_PATHWAYS_AND_PEOPLE },
-      ],
-      awaitRefetchQueries: true,
-      onCompleted: () => {
-        changeConnection(connectionsData[0] || {}) // ! if no more connections, pass empty object
-        setWhetherUnsavedChanges(false)
-      },
-      onError: window.alert,
-    }
-  )
+  const [deleteConnection] = useMutation(mutationDocs.delete, {
+    variables: {
+      input: { _id: stagedConnection._id },
+    },
+    refetchQueries,
+    awaitRefetchQueries: true,
+    onCompleted: () => {
+      selectConnectionId(_.isEmpty(connections) ? null : connections[0]._id)
+      setWhetherUnsavedChanges(false)
+    },
+    onError: window.alert,
+  })
 
   const deleteHandler = () => {
     if (
@@ -154,20 +90,6 @@ const ButtonCluster = ({
     ) {
       deleteConnection()
     }
-  }
-
-  const handleUpsertion = () => {
-    if (
-      areRequiredFieldsBlank(connectionData) &&
-      !exclusionSettings.isExcluded
-    ) {
-      window.alert(
-        `Please fill out the fields marked required OR check off "Exclude From Tool"`
-      )
-      return
-    }
-
-    upsert()
   }
 
   return (
@@ -188,7 +110,7 @@ const ButtonCluster = ({
         ) : (
           <Button
             type="secondary"
-            onClick={handleUpsertion}
+            onClick={upsert}
             color={Color.GREEN}
             buttonStyle={{ margin: `0 ${Spacing.S3}` }}
           >
@@ -228,13 +150,13 @@ const ButtonCluster = ({
 ButtonCluster.propTypes = {
   isNewConnectionBeingCreated: PropTypes.bool.isRequired,
   cancelHandler: PropTypes.func.isRequired,
-  connectionData: PropTypes.object.isRequired,
+  stagedConnection: PropTypes.object.isRequired,
   setWhetherUnsavedChanges: PropTypes.func.isRequired,
   setWhetherNewConnectionBeingCreated: PropTypes.func.isRequired,
-  changeConnection: PropTypes.func.isRequired,
-  connectionsData: PropTypes.array.isRequired,
+  selectConnectionId: PropTypes.func.isRequired,
+  connections: PropTypes.array.isRequired,
+  refetchQueries: PropTypes.array.isRequired,
+  mutationDocs: PropTypes.array.isRequired,
 }
-
-ButtonCluster.defaultProps = {}
 
 export default ButtonCluster