+++ src/frontend/components/BusinessObjectModal/shared/widget/mbmPayerConnectionComponents/ConnectionsList.js
@@ -1,6 +1,5 @@
 import React from 'react'
 import PropTypes from 'prop-types'
-import _ from 'lodash'
 import styled from '@emotion/styled'
 
 import ConnectionListItem from './ConnectionListItem'