+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Build out LBM basic CRUD and Account Overview [#886](https://github.com/pulse-analytics/wave-db-mgmt/pull/886)
   - Build out LBM types basic CRUD and power select [#887](https://github.com/pulse-analytics/wave-db-mgmt/pull/887)
   - Build out LBM <=> LBM types connections widget and CRUD [#888](https://github.com/pulse-analytics/wave-db-mgmt/pull/888)
+  - Build out LBM <=> Payer connections widget and CRUD [#892](https://github.com/pulse-analytics/wave-db-mgmt/pull/892)
+  - Build out LBM Payer Partnerships template table view [#893](https://github.com/pulse-analytics/wave-db-mgmt/pull/893)
 
 ## [4.2.0] - Started: 12-11-20